+++ tests/Rest/PostSubmissionControllerTest.php
@@ -19,14 +19,11 @@ class PostSubmissionControllerTest extends AbstractSubmissionControllerTestCase
     public function testAddSubmissionWithCreateFormSubmissionsPermission()
     {
         $form = $this->addForm(grantBasedSubmissionAuthorization: false); // test with available tags
-
         $this->authorizationTestEntityManager->addAuthorizationResourceAndActionGrant(
             AuthorizationService::FORM_RESOURCE_CLASS, $form->getIdentifier(),
             AuthorizationService::CREATE_SUBMISSIONS_FORM_ACTION, self::CURRENT_USER_IDENTIFIER);
 
-        $this->authorizationService->clearCaches();
-
-        $tags = [AbstractTestCase::TEST_AVAILABLE_TAGS[0]];
+        $tags = [];
         $submission = $this->postSubmission($form->getIdentifier(), '{"foo": "bar"}', tags: $tags);
         $this->assertEquals($submission->getForm()->getIdentifier(), $submission->getForm()->getIdentifier());
         $this->assertEquals($submission->getDataFeedElement(), $submission->getDataFeedElement());
@@ -42,19 +39,18 @@ class PostSubmissionControllerTest extends AbstractSubmissionControllerTestCase
         $this->assertEquals($tags, $gotSubmission->getTags());
         $this->assertEquals([], $gotSubmission->getGrantedActions());
 
-        $form = $this->addForm(grantBasedSubmissionAuthorization: true);
+        $this->authorizationService->clearCaches();
 
+        $form = $this->addForm(grantBasedSubmissionAuthorization: true);
         $this->authorizationTestEntityManager->addAuthorizationResourceAndActionGrant(
             AuthorizationService::FORM_RESOURCE_CLASS, $form->getIdentifier(),
             AuthorizationService::CREATE_SUBMISSIONS_FORM_ACTION, self::CURRENT_USER_IDENTIFIER);
 
-        $this->authorizationService->clearCaches();
-
-        $submission = $this->postSubmission($form->getIdentifier(), '{"foo": "bar"}', tags: $tags);
+        $submission = $this->postSubmission($form->getIdentifier(), '{"foo": "bar"}');
         $this->assertEquals($submission->getForm()->getIdentifier(), $submission->getForm()->getIdentifier());
         $this->assertEquals($submission->getDataFeedElement(), $submission->getDataFeedElement());
         $this->assertEquals(Submission::SUBMISSION_STATE_SUBMITTED, $submission->getSubmissionState());
-        $this->assertEquals($tags, $submission->getTags());
+        $this->assertEquals([], $submission->getTags());
         $this->assertEquals([], $submission->getGrantedActions());
 
         $gotSubmission = $this->getSubmission($submission->getIdentifier());
@@ -62,25 +58,24 @@ class PostSubmissionControllerTest extends AbstractSubmissionControllerTestCase
         $this->assertEquals($submission->getForm()->getIdentifier(), $gotSubmission->getForm()->getIdentifier());
         $this->assertEquals($submission->getDataFeedElement(), $gotSubmission->getDataFeedElement());
         $this->assertEquals(Submission::SUBMISSION_STATE_SUBMITTED, $gotSubmission->getSubmissionState());
-        $this->assertEquals($tags, $gotSubmission->getTags());
+        $this->assertEquals([], $gotSubmission->getTags());
         $this->assertEquals([], $gotSubmission->getGrantedActions());
     }
 
     public function testAddSubmissionWithManageFormPermissions(): void
     {
         $form = $this->addForm(grantBasedSubmissionAuthorization: false); // test with NO available tags
-
         $this->authorizationTestEntityManager->addAuthorizationResourceAndActionGrant(
             AuthorizationService::FORM_RESOURCE_CLASS, $form->getIdentifier(),
             AuthorizationService::MANAGE_ACTION, self::CURRENT_USER_IDENTIFIER);
 
-        $this->authorizationService->clearCaches();
-        $submission = $this->postSubmission($form->getIdentifier(), '{"foo": "bar"}');
+        $tags = [AbstractTestCase::TEST_AVAILABLE_TAGS[0]];
+        $submission = $this->postSubmission($form->getIdentifier(), '{"foo": "bar"}', tags: $tags);
         $this->assertEquals($submission->getIdentifier(), $this->getSubmission($submission->getIdentifier())->getIdentifier());
         $this->assertEquals($submission->getForm()->getIdentifier(), $submission->getForm()->getIdentifier());
         $this->assertEquals($submission->getDataFeedElement(), $submission->getDataFeedElement());
         $this->assertEquals(Submission::SUBMISSION_STATE_SUBMITTED, $submission->getSubmissionState());
-        $this->assertEquals([], $submission->getTags());
+        $this->assertEquals($tags, $submission->getTags());
         $this->assertEquals([AuthorizationService::MANAGE_ACTION], $submission->getGrantedActions());
 
         $form = $this->addForm(grantBasedSubmissionAuthorization: true);
@@ -230,25 +225,6 @@ class PostSubmissionControllerTest extends AbstractSubmissionControllerTestCase
         $this->assertEquals($submission->getIdentifier(), $gotSubmission->getIdentifier());
         $this->assertEquals($submission->getForm()->getIdentifier(), $gotSubmission->getForm()->getIdentifier());
         $this->assertEquals($submission->getDataFeedElement(), $gotSubmission->getDataFeedElement());
-    }
-
-    public function testAddSubmissionWithTags()
-    {
-        $form = $this->addForm();
-        $this->authorizationTestEntityManager->addAuthorizationResourceAndActionGrant(
-            AuthorizationService::FORM_RESOURCE_CLASS, $form->getIdentifier(),
-            AuthorizationService::CREATE_SUBMISSIONS_FORM_ACTION, self::CURRENT_USER_IDENTIFIER);
-
-        $this->authorizationService->clearCaches();
-        $submission = $this->postSubmission($form->getIdentifier(), '{"foo": "bar"}',
-            tags: [AbstractTestCase::TEST_AVAILABLE_TAGS[1]]);
-
-        $gotSubmission = $this->getSubmission($submission->getIdentifier());
-        $this->assertEquals($submission->getIdentifier(), $gotSubmission->getIdentifier());
-        $this->assertEquals($submission->getForm()->getIdentifier(), $gotSubmission->getForm()->getIdentifier());
-        $this->assertEquals($submission->getDataFeedElement(), $gotSubmission->getDataFeedElement());
-        $this->assertEquals(Submission::SUBMISSION_STATE_SUBMITTED, $gotSubmission->getSubmissionState());
-        $this->assertEquals([AbstractTestCase::TEST_AVAILABLE_TAGS[1]], $gotSubmission->getTags());
     }
 
     public function testAddSubmissionWithFiles()